twitter-clone-python/requirements.txt
mysql-connector-python~=8.0.25
prettytable~=2.1.0